requirements.txt
streamlit>=1.32.0
spacy>=3.7.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
pymupdf>=1.23.0
python-docx>=1.1.0
transformers>=4.38.0
torch>=2.2.0
numpy>=1.26.0
pandas>=2.2.0
plotly>=5.19.0
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0
pydantic>=2.6.0
pytest>=8.0.0
